Remove ancient legacy stuff from Gemspec
We seem to have a bunch of affordances in our `gemspec` for ancient
versions of rubygems.

A cursory look at this makes it clear that this wasn't actually maintained
(rspec specified 5 times, instead of the actual dependencies, for example),
and it's the year of our Lord 2021, we don't need this anymore.