Finish the switch to `rubyzip`
We switched from `zipruby` to `rubyzip` in 2018
(commit d65d5a3166edd2b152070207536b6f21a4e9b41d)

However, we forgot to change the `gemspec` in the proces (oopsie), so
apps that use this gem are still installing `zipruby`, which hasn't been
touched in 10 years, and doesn't compile today without magic compiler
flags.

This finishes that upgrade.

I chose version `1.3.0` because our main app happens to be using another
gem which requires `rubyzip`, and it specifies `< 2.0.0`, and I want to
limit the rabbit hole i'm digging for myself here. But ideally we'd move
this to the latest version.